Initial commit.
This is a project I started years ago when Amato Opera Theatre was still going.
Irene Kim had asked me to come up with software to replace the paper-based
system they had for scheduling performances. (c 2007)   I never got to deploy this
as Amato closed before it was finished.   I intend to start this back up, only
as a Django project, rather than java, and with greater scope.